[IMP] mail: unskip "composer: send button is disabled if attachment upload is not finished"
