+++ addons/mail/static/tests/qunit_suite_tests/components/composer_tests.js
@@ -55,79 +55,6 @@ QUnit.module("mail", (hooks) => {
         });
 
         QUnit.skipRefactoring(
-            "composer: send button is disabled if attachment upload is not finished",
-            async function (assert) {
-                assert.expect(8);
-
-                const pyEnv = await startServer();
-                const attachmentUploadedPromise = makeTestPromise();
-                const mailChannelId1 = pyEnv["mail.channel"].create({});
-                const { messaging, openDiscuss } = await start({
-                    discuss: {
-                        context: { active_id: mailChannelId1 },
-                    },
-                    async mockRPC(route) {
-                        if (route === "/mail/attachment/upload") {
-                            await attachmentUploadedPromise;
-                        }
-                    },
-                });
-                await openDiscuss();
-                const file = await createFile({
-                    content: "hello, world",
-                    contentType: "text/plain",
-                    name: "text.txt",
-                });
-                await afterNextRender(() =>
-                    inputFiles(messaging.discuss.threadView.composerView.fileUploader.fileInput, [
-                        file,
-                    ])
-                );
-                assert.containsOnce(
-                    document.body,
-                    ".o_AttachmentCard",
-                    "should have an attachment after a file has been input"
-                );
-                assert.containsOnce(
-                    document.body,
-                    ".o_AttachmentCard.o-isUploading",
-                    "attachment displayed is being uploaded"
-                );
-                assert.containsOnce(
-                    document.body,
-                    ".o-mail-composer-send-button",
-                    "composer send button should be displayed"
-                );
-                assert.ok(
-                    !!document.querySelector(".o-mail-composer-send-button").attributes.disabled,
-                    "composer send button should be disabled as attachment is not yet uploaded"
-                );
-
-                // simulates attachment finishes uploading
-                await afterNextRender(() => attachmentUploadedPromise.resolve());
-                assert.containsOnce(
-                    document.body,
-                    ".o_AttachmentCard",
-                    "should have only one attachment"
-                );
-                assert.containsNone(
-                    document.body,
-                    ".o_AttachmentCard.o-isUploading",
-                    "attachment displayed should be uploaded"
-                );
-                assert.containsOnce(
-                    document.body,
-                    ".o-mail-composer-send-button",
-                    "composer send button should still be present"
-                );
-                assert.ok(
-                    !document.querySelector(".o-mail-composer-send-button").attributes.disabled,
-                    "composer send button should be enabled as attachment is now uploaded"
-                );
-            }
-        );
-
-        QUnit.skipRefactoring(
             "remove an attachment from composer does not need any confirmation",
             async function (assert) {
                 assert.expect(3);
